Created cable droppage
The cable is in it's basic idea, let me know if you want to change it.
It is strange because it believes it is in 3d... So we either have to
cope with it or find a workaround

Basic movement was added, along with a camera to follow it.

Please ignore the sprite image... It was just a random thing I found and
it works