I tried to implement pulling a rope...
It didn't work out so well. It is currently commented out, but line
renderer is not meant to calculate something like that. In order to
render that to full potential, it has to loop through each individual
rope segment, and poke around with it. Got incredibly laggy once I got
to around 25 segments. I say we don't implement it and find a
workaround.

Also cleaned up spare code.