Updated fader, fixed holding system
Fader now can start with any size. With the current size, it was
becoming incredibly big and in the way, so the script now increases the
size when it blacks out. However, some sprites still appear in front of
the fader. Watch out for that.

I made the item pickup system take in 3 creates with the "Loot" script
inside of them. It is a public value, so it can be increased through
upgrades.