Bug fies, updates.
-Made thruster sprites smaller
-fixed bug with exceptions being thrown as any object enters the trigger
of the ship
-crane now can be destroyed with lasers
-crane rope can be severed and not be retrieved
-added an icon displayed whenever close to the ship, telling the player
to return
-added a system to send a message if any script includes certain methods
and a message to send, along with an action to take if message is sent
-updated prefabs
-changed itemholder methods for simpler code and easier returning
-added another warning to show if your crane is destroyed
-if your crane is destroyed, the upgrade screen will know and will set
it to 1 after ten easy payments of $9.99!
-Messed around with GUITextures, which are really weird. Working on
that.

There is a strange bug dealing with health and your health turning into
infinity, going to take a lot of playtesting to get to that.