Finalized door class and button class
In order to use a door, all you have to do is set the DoorOpen and
DoorClose where you want the door to open to and close to, respectively.
Then have as many doors as you want be a child of a button, which will
then control them.

Only bug: door closing on you does not kill you.